Enable and disable of the exporter, supporter and customs in the admin panel

+++ admin/customsview.php
@@ -22,7 +22,7 @@ else
     exit;
 }
 
-$custom_sql="SELECT ci.*, `countries`.`name` AS country_name, cities.city_name, states.state_name, terminals.terminals_name, ports.ports_name
+$custom_sql="SELECT ci.*, `countries`.`name` AS country_name, cities.city_name, states.state_name, terminals.terminals_name, ports.ports_name, users.delete_status
 FROM `customs_info` AS `ci`
 LEFT JOIN `users` ON users.id = ci.user_id
 LEFT JOIN `countries` ON countries.id = ci.country
@@ -30,7 +30,7 @@ LEFT JOIN `cities` ON cities.id = ci.city
 LEFT JOIN `states` ON states.id = ci.state
 LEFT JOIN `ports` ON ports.id = ci.port
 LEFT JOIN `terminals` ON terminals.id = ci.terminal
-WHERE `users`.delete_status = 1 and `ci`.id = $customs_id";
+WHERE `ci`.id = $customs_id";
 $custom_exe=mysql_query($custom_sql);
 $custom_cnt=@mysql_num_rows($custom_exe);
 $custom_fet=mysql_fetch_array($custom_exe);
@@ -123,6 +123,21 @@ $custom_fet=mysql_fetch_array($custom_exe);
                                         <label class="col-sm-3 control-label">Terminal</label>
                                         <div class="col-sm-9"><div class=""><?php echo $custom_fet['terminals_name']; ?></div></div>
                                     </div>
+                                    <div class="form-group col-md-12">
+                                        <label class="col-sm-3 control-label">Status</label>
+                                        <div class="col-sm-9"><div class="">
+                                                <?php if($custom_fet['delete_status'] == 1)
+                                                {?>
+                                                    <button type="button" class="btn btn-success btn-xs">Active</button>
+                                                <?php
+                                                }
+                                                else{
+                                                    ?>
+                                                    <button type="button" class="btn btn-danger btn-xs">Inactive</button>
+                                                <?php
+                                                }?>
+                                            </div></div>
+                                    </div>
                                 </div>
                                 <div class="col-md-1"></div>
                             </div><!-- /.box-body -->
@@ -143,7 +158,21 @@ $custom_fet=mysql_fetch_array($custom_exe);
                             <div class="form-group col-md-12">
                                 <a href="customslist.php"><button type="submit" class="btn btn-warning col-md-12" style="margin-bottom:10px;" >Back to Customs List</button></a>
                                 <a href="customsedit.php?customs_id=<?php echo $custom_fet['id']; ?>"><button type="button" class="btn btn-danger col-md-12" style="margin-bottom:10px;"><i class="fa fa-pencil"></i> Edit</button></a>
-                                <a href="customs-delete.php?delete=1&customs_id=<?php echo $custom_fet['user_id']; ?>" onclick="return confirm('Are you sure you want to disable this item?');"><button type="button" class="btn btn-warning col-md-12"><i class="fa fa-trash-o"></i> Disable</button></a>
+
+                                <?php
+                                if($custom_fet['delete_status'] == 1){
+                                    ?>
+                                    <a href="customs-delete.php?delete=1&customs_id=<?php echo $custom_fet['user_id']; ?>" onclick="return confirm('Are you sure you want to disable this item?');"><button type="button" class="btn btn-warning col-md-12"><i class="fa fa-trash-o"></i> Disable</button></a>
+                                <?php
+                                }
+                                else{
+                                    ?>
+                                    <a href="customs-delete.php?enable=1&customs_id=<?php echo $custom_fet['user_id']; ?>" onclick="return confirm('Are you sure you want to enable this item?');"><button type="button" class="btn btn-success col-md-12"><i class="ion ion-pie-graph"></i> Enable</button></a>
+                                <?php
+                                }
+                                ?>
+
+
                             </div>
                         </div><!-- /.box-body -->
                     </div><!-- /.box -->
